src/nl/rug/escher/entities/Study.java
package nl.rug.escher.entities;

import java.util.List;

import com.jgoodies.binding.beans.Model;

public class Study extends Model {
	private String d_id;
	private List<Endpoint> d_endpoints;
	
	public List<Endpoint> getEndpoints() {
		return d_endpoints;
	}

	public void setEndpoints(List<Endpoint> endpoints) {
		List<Endpoint> oldVal = d_endpoints;
		d_endpoints = endpoints;
		firePropertyChange(PROPERTY_ENDPOINTS, oldVal, d_endpoints);
	}

	public final static String PROPERTY_ID = "id";
	public final static String PROPERTY_ENDPOINTS = "endpoints";

	public String getId() {
		return d_id;
	}

	public void setId(String id) {
		String oldVal = d_id;
		d_id = id;
		firePropertyChange(PROPERTY_ID, oldVal, d_id);
	}
	
	public String toString() {
		return getId();
	}
}
